+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGELOG
 
+## v0.1.2 (2022-03-11)
+  * install zombie wrapper before run
+  * uncomment module, fix hexdocs
+
 ## v0.1.0 (2022-03-07)
   * Blatant rip off/trim of phoenixframework/tailwind